student-submissions/ehudab/week-01/requirements.txt
# Requirements for Week 1 AI Agents (OpenRouter + LangChain)
python-dotenv
langchain
langchain_openai
langchain_community
openai 